backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
jinja2
requests
